Add files via upload
This is an A search for the available AI game taxi (taxi-v2). 
A search is used to find the best path from the game's starting state. 
The grid of the state space is 5x5. 
It transforms the number representation into a coordinate system, with (1,1) at the bottom left and (5,5) at the top right.
First order logic is used in picking up function.
Manhatten Distance is used for the heuristic function.